requirements.txt
DateTime==4.3
greenlet==1.1.2
logger==1.4
numpy==1.22.0
pandas==1.3.5
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.29
zope.interface==5.4.0
